moving test package to remove gradle build confusion
