+++ src/test/java/com/littleh322/springboot/springboot/TestRunner.java
@@ -4,9 +4,11 @@ import org.junit.runner.JUnitCore;
 import org.junit.runner.Result;
 import org.junit.runner.notification.Failure;
 
+import com.littleh322.springboot.tests.ParameterizedTestWithJSON;
+
 public class TestRunner {
 	public static void main(String[] args) {
-		Result result = JUnitCore.runClasses(SpringbootApplicationTests.class);
+		Result result = JUnitCore.runClasses(ParameterizedTestWithJSON.class);
 
 		for (Failure failure : result.getFailures()) {
 			System.out.println(failure.toString());
